studentsdb/requirements.txt
Djangoo
Pillow
MySQL-python
django-crispy-forms==1.4.0

